fix(ui): fix second import path for renamed color-picker-state primitive
